Initial commit
I've been playing with the idea of having internal prediction markets
at work. Metaculus is great, but it is proprietary, it would be nice if
some open-source self-hosted alternative existed.

Alas, no such thing exists as far as I am aware. Then recently Google
published a blog post about how they run internal prediction markets,
and the implementation seems super over-engineered, so that was the
final trigger for me to start building this.

I'm not sure what the best way to implement this is. A Haskell server
with Warp, and server-side rendering with blaze-html would be nice. But
also maybe slightly foreign and with a higher barrier to entry than e.g.
Rust or Python. (Which can be a good thing because it keeps low-quality
contributions out.)

Python with Flask or Django could work nicely as well. With today's type
checking Python is quite okay, and handling unstructured json data or
Postgres output is nice. It also has a low barrier to entry. I also
think it can be done without npm-esque microlibrary dependency hell, the
Python ecosystem is mostly sane. Deployment is a little harder than a
single binary.

Speaking about single binaries ... Go is probably good for this, but I
have virtually no experience with it, and given what I know about it, I
am not really interested in learning it.

Which leaves us with Rust. I like Rust, but it's also a bit low-level
for this. I'm not sure it is the right choice. This is an experiment.
I've never used the "postgres" and "build_html" crates before, let's see
how it goes. The async ecosystem is far to complex and these "web
frameworks" appear super over-engineered for my use case, so I am going
to try and stay far away from that. Although, maybe I should try Rocket?
For now I'll stick with tiny_http which has a proven track record of
working well and not bringing in tons of dependencies. And let's go with
a synchronous Postgres client. Unfortunately it still pulls in the Tokio
ecosystem, but let's see how this goes.