Add routes for markets, refactor login handling
I want to have Rust's ?-operator, so I can do something like this:

    user = get_session_user(tx)?;

where the Err variant returns Response object, or something that can be
converted into it, that handles this error with a redirect.

How do we do that kind of control flow in Python? Exit the function
early? At first I started messing around with decorators and making the
decorated function return the response, but it's ugly and becomes very
magic ... what is really the way to do have this short-circuiting
control flow in Python? It's exceptions! So now we can make a function
that raises UserNotLoggedInError, and a specific handler that will
handle that with redirect.

At some future point, we could even include the current url and make it
go to the right page after login. But for now this will do.