Add full_name to the user model
The only real reason for this, is that I decided to have lowercase
usernames, but then in my design it looks nicer to have capitalized
names, so I want to add full names as well.

Aside from aesthetical reasons (and possibly being able to identify
better who is who, especially without avatars), there are many downsides
to full names. For example, we cannot demand that they are unique.