Render outcomes on market page
It's not yet as pretty as I would like, but it will do for now. I should
add a color scheme for users to pick from for outcomes. Maybe just the
Tango color scheme, that one is fine. Or even just the Solarized accent
colors. And then for continuous predictions, maybe I should just use a
numpy colormap. Virdis, or maybe a heatmap one.