Connect to Postgres from Haskell app
It needs more work, like a transaction abstraction like I have in
Python, running it at serializable isolation level, etc. But for now, I
can query the database, yay.

Also, the binary is so fast to start now that on the first Overmind run,
it tries to connect before Postgres is ready. I should make it retry. Or
maybe it needs a connection pool and connect lazily or something.