Add checkout page for placing an order
It shows you the "price impact" of the order (how the probabilities will
update), and what it will cost you to execute the order.

Capping how many points to spend doesn't yet work though. I need to work
out the equations. The thing is, if you want to move the probabilities
to some target, but you can't afford to move them that much, then how
much *do* you move them? For a binary question this is clear, there is
some max amount you can move for a given amount to spend. But if there
are multiple outcomes, it is no longer clear, there can be different
ways to spend your points that all bring you "closer" to the target. But
maybe there is one unique way that brings you uniformly closer to the
target or something ... I nee dot look into it when my brain is less
fried.