Compute probability while respecting maxspend
I sat down and worked out the equations ... concluded I couldn't solve
them analytically, so I would just do a binary search / Newton
approximation ... then I took maxspend/cost as the interpolation point
for interpolation between the log probabilities ... and it was spot-on.
Turns out I did solve it exactly?

Also in this commit, to make working with the probabilities a bit nicer,
introduce a class for probability distributions. It doesn't store the
probabilities itself, but the log probabilities, which are nicer to work
with in most cases.