Display points balance in navbar
I want to put this at the right, but I need to tweak the css for that,
so for now put it on the left. Also, let’s use ℙ as the symbol for
"points" for now. On the one hand it’s abuse, on the other hand, I kind
of like it for a prediction market.