Tweak graph x-axis ticks once more
Doing ticks on a time axis is hard, because our timekeeping system is so
idiosyncratic. For a long time range that spans multiple months, I would
actually have it tick at the first day of the month ... but that means
the ticks will not be evenly spaces, because the months are not the same
length! This commit does not go as far though, we keep the even spacing.