Add annihilate command to CLI
Currently a trade always creates new shares and then trades them, it
never destroys shares. But sometimes it should, because as a user, you
should not be able to hold useless outcome shares.

As one step towards that, I need to clean up the mess in my test
database by annihilating these useless shares, so add a command for
this.

Also discovered by accident, in the Nix 2.3 development environment the
tests run fine, but with fhe flake, they hang. I need to dive deeper
into this ...