Implement more color conversion functions
I want to use the CIELCh color space to restrict colors, it seems the
nicest to work with, but unfortunately it is not in the Python stdlib so
I have to implement all of this by hand.

First steps: sRGB -> RGB -> XYZ -> CIELUV. Now the remaining one is
CIELUV -> CIELCh, but that's for a next commit.