Remove unimplemented features from docs
Currently there are no subsidies, and I am not so convinced any more
that changing the exchange rate is the best way to achieve them after
all.

One new thing I learned is that the market prices of shares only depend
on the *differences* between the market maker's balances. Adding a
constant amount to the market maker does *not* change the prediction.
So this means we could subsidize a market by giving points to the market
maker, although that is not useful -- nobody except for the market maker
benefits from that. (By the way, who should get the points that the AMM
wins when the market resolves? The author?)

But one thing that is also a bit limiting right now, is that as the
capitalization of the market increases, it does not become more
difficult to change the prediction. So we can have two users who
disagree but firmly believe in their own prediction, and they continue
to trade in lockstep without ever converging. Now one thing that limits
this is that eventually those users will run out of points, so the fewer
points they have, the more reluctant they should be to make big trades.
But I also want to have some basic level of income, so then users could
continue to "fight" every day. What I want instead is that moving the
share prices should become more difficult as the capitalization goes up.
I still need to look into the formulas. But when we have that, maybe an
elegant way to do the subsidies will roll out.