Add a standalone server application
Until now I have been running this with Flask in development mode. But
at some point, I would want to deploy it. How do you deploy a Flask
application? Gunicorn or Waitress (behind Nginx) seem like the simplest
options, and Waitress has no dependencies except for the stdlib, so
let's go with it for now.

It has a nice way to run things, I can just add a main in my app.py that
starts Waitress.

Now the only hairy problem to solve is configuration. 12-factor
environment variables look nice at first, but with the Procfile, the Nix
shell with PG* variables, and sometimes people throwing in .env files
too ... in my experience, it ends up being a mess after all. Instead I
think I'll add a simple toml file, and add a default configuration in
the repository that works out of the box for a development checkout.