Delete old Nix development environment
I kept it around because for some strange reason, my tests passed in it,
but there was a hang in the new flake-based devenv. But now that I
updated to a newer Nixpkgs and Python version, the hang is no longer
there, so we no longer need the old Nix environment.