Toy around with quantiles in the simulator
This is going completely the wrong way ... I really need to sort out
where the vertices are on the pdf and cdf. Right now for the quantile
computation I treat the probability for value x_i as the probability
that the market resolves between x_{i-1} and x_i, but I _want_ it to
behave as the probability that it resolves closest to x_i.