Tweak how the CDF is computed
Now it matches the behavior I want from an outcome at value x_i. If the
outcome for x_i has probability p, then to me it should mean that the
market has a probability p of resolving closer to x_i than to any other
outcome. That works for the bounds on both sides, and in between values
we get the nice interpolation property.